Starting to work on this app
The goal is to make a simple question application that a user can specify a URL or a file that contains the list of questions in a JSON format. I'm using the list of questions provided to me from a course I've previously taken initially and will be writing my own for an example in the future.